refactor(cmd): thin cmd/ via internal/cli + breaker.Registry; atomic SIGHUP
Closes #609 and #615.

cmd/s3-orchestrator drops from 1,770 to 217 LoC. Subcommand bodies relocate
to internal/cli/{adminctl,initcmd,synccmd,serve}; cmd/ now only routes
os.Args[1] to thin shims.

internal/breaker/registry.go owns the set of breakers the watchdog sweeps.
ProvideBreakerRegistry assembles the database CB + per-backend CBs at DI
construction time, so circuitBreakerWatchdog no longer type-asserts
backends.

reloadConfig becomes validate-then-apply: a failing LoadConfig aborts
without touching cfgPtr; on success applyReload performs the swap. Dead
ProvideLogBuffer dropped. Test covers the rollback path.

Subcommand additions:
- s3-orchestrator admin reconcile [-backend NAME] now wraps the existing
  /admin/api/reconcile endpoint that the docs already described.

Documentation refresh across README, docs/, web/content/diagrams/, and
web/content/godoc/breaker.md to track the new layout, retire stale
MetadataStore mentions, and bump version examples.
