+++ .version
@@ -1,1 +1,1 @@
-v0.41.6
+v0.41.7
